moved the ribboncontroller to the ribbonnodes package where it belongs, improved the genomicfeature class with a correct implementation of overlap checking and better precondition checking
