+++ backEnd/src/main/java/controller/Controller.java
@@ -11,6 +11,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import phylogenetictree.PhylogeneticTree;
+import ribbonnodes.RibbonController;
 import ribbonnodes.RibbonNode;
 
 /**
